requirements.txt
streamlit>=1.24.0
requests>=2.31.0 
openinference-semantic-conventions
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
arize-phoenix